% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conventional_statations_hidroweb.R
\name{conventional}
\alias{conventional}
\title{Download data from convetional stations}
\usage{
conventional(code_hidroweb, var)
}
\arguments{
\item{code_hidroweb}{integer. HidroWeb code for conventional stations}

\item{var}{character. Valid variable names are "vazoes", "cotas", "chuvas"}
}
\description{
Download data from convetional stations
}
\examples{
code_hidroweb <- 75890000
var = "cotas"
conventional(code_hidroweb,var)
}
