% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telemetry_hidroweb.R
\name{list_of_telemetric_stations}
\alias{list_of_telemetric_stations}
\title{List of Telemetric stations}
\usage{
list_of_telemetric_stations(Estacoes = 0, origem = 0)
}
\arguments{
\item{Estacoes}{integer. 0 - active stations and 1 - in maintanance}

\item{origem}{integer.  0-Todas, 1-ANA/INPE, 2-ANA/SIVAM, 3-RES_CONJ_03, 4-CotaOnline, 5-Projetos Especiais}
}
\description{
List of Telemetric stations
}
