% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huff_storms.R
\name{rain_events}
\alias{rain_events}
\title{Filtering of rainfall events from sub-daily time series}
\usage{
rain_events(date, depth, MIT = 6, Dmin = 13, Thres = 0.5)
}
\arguments{
\item{date}{A vector containing date of the rainfall time series. The date must be as POSIXct class.}

\item{depth}{A vector containing the precipitation depth time series in mm.}

\item{Dmin}{The minimum}
}
\description{
Filtering of rainfall events from sub-daily time series
}
\examples{
MIT = 6 #
Dmin = 13 # minimum event depth
Thres = 0.5
}
