% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conventional_hidro.R,
%   R/conventional_statations_hidroweb.R
\name{conventional}
\alias{conventional}
\title{Donwload data from conventional stations}
\usage{
conventional(code_hidroweb, var)

conventional(code_hidroweb, var)
}
\arguments{
\item{code_hidroweb}{integer. HidroWeb code for conventional stations}

\item{var}{character. Valid variable names are "vazoes", "cotas", "chuvas"}

\item{codEstacao}{integer. Code of rain or flow station}

\item{dataInicio}{character. Date of the beginning of the observation in the format 'day/month/Year'}

\item{dataFim}{character. Date of the ending of the observation in the format 'day/month/Year'}

\item{tipoDados}{integer. 1-Cotas, 2-Chuvas ou 3-Vazões for stage, rain or flow, respectively}

\item{nivelConsistencia}{integer. 1-Bruto ou 2-Consistido}
}
\description{
Donwload data from conventional stations

Download data from convetional stations
}
\examples{
code_hidroweb <- 75890000
var = "cotas"
conventional(code_hidroweb,var)
}
