% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conventional_hidro.R
\name{conventional}
\alias{conventional}
\title{Donwload data from conventional stations}
\usage{
conventional(CodEstacao, DataInicio, DataFim, tipoDados, nivelConsistencia)
}
\arguments{
\item{tipoDados}{integer. 1-Cotas, 2-Chuvas ou 3-Vazões for stage, rain or flow, respectively}

\item{nivelConsistencia}{integer. 1-Bruto ou 2-Consistido}

\item{codEstacao}{integer. Code of rain or flow station}

\item{dataInicio}{character. Date of the beginning of the observation in the format 'day/month/Year'}

\item{dataFim}{character. Date of the ending of the observation in the format 'day/month/Year'}
}
\description{
Donwload data from conventional stations
}
